add authInfo method to ChannelFilterCallbacks; move AuthInfo to separate header
