+++ test/extensions/filters/network/ssh/test_mocks.h
@@ -121,6 +121,7 @@ public:
   MOCK_METHOD(std::optional<std::string_view>, channelType, (), (const));
   MOCK_METHOD(Stats::Scope&, scope, (), (const));
   MOCK_METHOD(stream_id_t, streamId, (), (const));
+  MOCK_METHOD(const AuthInfo&, authInfo, (), (const));
   MOCK_METHOD(bool, interruptChannel, (absl::Status));
 };
